Split active tabs out into a new file
